# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Model Computer Vision API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/similarity/embed": {
      "post": {
        "tags": [
          "Similarity"
        ],
        "summary": "Compute an image embedding",
        "description": "Turn an image into an embedding vector without storing anything.\n\nUse this when you keep your own vector store. The returned vector is\nnormalised to unit length, so the cosine similarity between two of them is\nsimply their dot product.",
        "operationId": "embed_api_v1_similarity_embed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            }
          },
          "400": {
            "description": "The image could not be decoded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The image exceeds the size limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported image format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for your tier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The embedding model is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/similarity/index": {
      "post": {
        "tags": [
          "Similarity"
        ],
        "summary": "Add an image to the similarity index",
        "description": "Embed an image and store it so future searches can find it.\n\nAttach a ``label`` and arbitrary ``metadata``; both are returned with any\nsearch hit, so the caller does not need a second lookup to find out what\nwas matched.",
        "operationId": "index_image_api_v1_similarity_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexResponse"
                }
              }
            }
          },
          "400": {
            "description": "The image could not be decoded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The image exceeds the size limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported image format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for your tier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The embedding model is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/similarity/search": {
      "post": {
        "tags": [
          "Similarity"
        ],
        "summary": "Find visually similar images",
        "description": "Find the indexed images most similar to the supplied image.\n\nScores are cosine similarities: ``1.0`` is an identical direction, ``0.0``\nunrelated, ``-1.0`` opposite. In practice, near-duplicates score above\n0.95 and \"same kind of thing\" lands somewhere around 0.7-0.9 \u2014 the right\n``min_similarity`` depends on your images and is worth tuning on real data\nrather than guessing.\n\nAn empty ``results`` list with ``index_size: 0`` means nothing has been\nindexed yet, not that nothing matched.",
        "operationId": "search_api_v1_similarity_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimilarityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarityResponse"
                }
              }
            }
          },
          "400": {
            "description": "The image could not be decoded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The image exceeds the size limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported image format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for your tier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The embedding model is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/similarity/upload": {
      "post": {
        "tags": [
          "Similarity"
        ],
        "summary": "Find similar images for an uploaded file (multipart)",
        "description": "Find similar images for a multipart file upload.",
        "operationId": "search_upload_api_v1_similarity_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_search_upload_api_v1_similarity_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarityResponse"
                }
              }
            }
          },
          "400": {
            "description": "The image could not be decoded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The image exceeds the size limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported image format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for your tier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The embedding model is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmbeddingRequest": {
        "properties": {
          "model_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[a-zA-Z0-9_\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "Registry model name. Defaults to the task's default model."
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "pattern": "^[a-zA-Z0-9_.\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version",
            "description": "Registry version, or 'latest'. Defaults to 'latest'."
          },
          "runtime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeFormat"
              },
              {
                "type": "null"
              }
            ],
            "description": "Force a specific runtime format. Defaults to the server's preferred runtime."
          },
          "image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Base64",
            "description": "Raw image bytes, base64-encoded. A `data:` URI prefix is accepted and stripped."
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Publicly reachable HTTPS URL to fetch the image from."
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id",
            "description": "Caller-supplied identifier echoed back in the response, useful for batches."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EmbeddingRequest",
        "description": "Body for ``POST /api/v1/similarity/embed`` (vector only, no search)."
      },
      "EmbeddingResponse": {
        "properties": {
          "model": {
            "$ref": "#/components/schemas/ModelInfo"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingInfo"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id",
            "description": "Quote this id in any support request."
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "True when this result was served from Redis instead of re-running the model.",
            "default": false
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id",
            "description": "Echo of the caller-supplied image_id, if any."
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "description": "True when the primary model was unavailable and a fallback model or runtime served the request instead.",
            "default": false
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Non-fatal notes, e.g. 'image was resized' or 'fell back to torch runtime'."
          },
          "embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Embedding",
            "description": "L2-normalised embedding vector."
          },
          "dimension": {
            "type": "integer",
            "title": "Dimension",
            "description": "Length of the embedding vector."
          }
        },
        "type": "object",
        "required": [
          "model",
          "timing",
          "correlation_id",
          "embedding",
          "dimension"
        ],
        "title": "EmbeddingResponse",
        "description": "Response for ``POST /api/v1/similarity/embed``."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The single error shape every failing endpoint returns."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndexImageRequest": {
        "properties": {
          "model_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[a-zA-Z0-9_\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "Registry model name. Defaults to the task's default model."
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "pattern": "^[a-zA-Z0-9_.\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version",
            "description": "Registry version, or 'latest'. Defaults to 'latest'."
          },
          "runtime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeFormat"
              },
              {
                "type": "null"
              }
            ],
            "description": "Force a specific runtime format. Defaults to the server's preferred runtime."
          },
          "image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Base64",
            "description": "Raw image bytes, base64-encoded. A `data:` URI prefix is accepted and stripped."
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Publicly reachable HTTPS URL to fetch the image from."
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id",
            "description": "Caller-supplied identifier echoed back in the response, useful for batches."
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Human-readable label stored alongside the vector."
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Arbitrary JSON metadata returned with search hits."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "IndexImageRequest",
        "description": "Body for ``POST /api/v1/similarity/index`` \u2014 add an image to the index."
      },
      "IndexResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identifier assigned to the newly indexed image."
          },
          "index_size": {
            "type": "integer",
            "title": "Index Size",
            "description": "Index size after the insert."
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "index_size",
          "correlation_id"
        ],
        "title": "IndexResponse",
        "description": "Response for ``POST /api/v1/similarity/index``."
      },
...
      "RuntimeFormat": {
        "type": "string",
        "enum": [
          "torch",
          "torch_int8",
          "onnx",
          "onnx_int8",
          "tensorrt"
        ],
        "title": "RuntimeFormat",
        "description": "Which compiled form of a model is executing."
      },
...
      "SimilarityRequest": {
        "properties": {
          "model_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[a-zA-Z0-9_\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "Registry model name. Defaults to the task's default model."
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "pattern": "^[a-zA-Z0-9_.\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version",
            "description": "Registry version, or 'latest'. Defaults to 'latest'."
          },
          "runtime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeFormat"
              },
              {
                "type": "null"
              }
            ],
            "description": "Force a specific runtime format. Defaults to the server's preferred runtime."
          },
          "image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Base64",
            "description": "Raw image bytes, base64-encoded. A `data:` URI prefix is accepted and stripped."
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Publicly reachable HTTPS URL to fetch the image from."
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id",
            "description": "Caller-supplied identifier echoed back in the response, useful for batches."
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "How many nearest neighbours to return.",
            "default": 10
          },
          "min_similarity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Min Similarity",
            "description": "Drop neighbours scoring below this cosine similarity.",
            "default": 0.0
          },
          "include_embedding": {
            "type": "boolean",
            "title": "Include Embedding",
            "description": "Also return the query image's raw embedding vector (large).",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SimilarityRequest",
        "description": "Body for ``POST /api/v1/similarity/search``."
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key issued to your account. Determines your tier."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT access token."
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
}